fix: turned CreateContactActivity to CreateContactFragment
