+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.CreateContactActivity" />
         <activity android:name=".activity.ModifyContactActivity" />
     </application>
